Rebase with and fix background for schedule
