+++ autoscheduler/frontend/templates/index.html
@@ -19,8 +19,8 @@
 
         <!-- Dependencies -->
         <!-- I think we need these? -->
-        <script crossorigin src="https://unpkg.com/react@16/umd/react.development.js"></script>
-        <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
+        <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react/16.12.0/umd/react.development.js"></script>
+        <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/16.11.0/umd/react-dom.development.js"></script>
 
         {% load static %}
         <script src="{% static "main.js" %}"></script>
